Removed deploy step and added variable test.
Probably going to change the Continuous Deployment methodology, so I removed that, at least until I've finalised plans for that.

To load the APIKey necessary for the Discord Bot to function into the build I plan on using DevOps Variable Groups, and this is testing to see if that works like expected.